Write functions to wait for atrial sensor signal
Wrote wait_for_atrial_sensor_high() and wait_for_atrial_sensor_low()
functions. These functions wait for a signal from the heart's atrial
sensor, represented by a button, to tell the pacemaker when the aitria
have contracted and when they've released. I added the functions into
the loop section of the 'main' function and verified that the button
and LEDs function as expected.